# src/main/resources/swagger.yaml
openapi: 3.0.0
info:
  title: CoActivity API
  description: API test
  version: 1.0.0

#servers:
#  - url: 'http://coactivity_app:8080/api/v1'
...
paths:
...
  /users/{id}/profile:
    get:
      tags:
        - Users
      summary: Войти в свой профиль
      operationId: getUserProfile
      parameters:
        - name: id
          in: path
          required: true
          description: ID текущего пользователя
          schema:
            type: integer
      responses:
        '200':
          description: Данные пользователя успешно получены
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/User'
        '404':
          description: Пользователь не найден
  /users/{id}/profile/logout:
    post:
      tags:
        - Users
      summary: Сменить профиль (Выход)
      description: Пользователь нажимает "Сменить профиль". Текущая сессия завершается.
      operationId: changeUserProfile
      parameters:
        - name: id
          in: path
          required: true
          description: ID пользователя, который выходит из системы
          schema:
            type: integer
      responses:
        '200':
          description: Успешный выход. Далее клиент должен перенаправить на Login/Register.
        '404':
          description: Пользователь не найден
  /users/{id}/profile/personal-info:
    get:
      tags:
        - Users
      summary: Просмотр персональной информации пользователя
      operationId: getPersonalInfo
      parameters:
        - name: id
          in: path
          required: true
          description: ID пользователя
          schema:
            type: integer
      responses:
        '200':
          description: Персональная информация успешно получена
          content:
            application/json:
              schema:
                type: object
                properties:
                  name:
                    type: string
                  email:
                    type: string
                  phone:
                    type: string
                  address:
                    type: string
        '404':
          description: Пользователь не найден
  /users/{id}/profile/personal-info/name:
    put:
      tags:
        - Users
      summary: Изменить имя пользователя
      operationId: updateUserName
      parameters:
        - name: id
          in: path
          required: true
          description: ID пользователя
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                name:
                  type: string
      responses:
        '200':
          description: Имя пользователя успешно обновлено
        '400':
          description: Некорректные данные
        '404':
          description: Пользователь не найден
  /users/{id}/profile/personal-info/email:
    put:
      tags:
        - Users
      summary: Изменить email пользователя
      operationId: updateUserEmail
      parameters:
        - name: id
          in: path
          required: true
          description: ID пользователя
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                email:
                  type: string
      responses:
        '200':
          description: Email пользователя успешно обновлен
        '400':
          description: Некорректные данные
        '404':
          description: Пользователь не найден
  /users/{id}/profile/personal-info/phone:
    put:
      tags:
        - Users
      summary: Изменить телефон пользователя
      operationId: updateUserPhone
      parameters:
        - name: id
          in: path
          required: true
          description: ID пользователя
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                phone:
                  type: string
      responses:
        '200':
          description: Телефон пользователя успешно обновлен
        '400':
          description: Некорректные данные
        '404':
          description: Пользователь не найден
  /users/{id}/profile/personal-info/address:
    put:
      tags:
        - Users
      summary: Изменить адрес пользователя
      operationId: updateUserAddress
      parameters:
        - name: id
          in: path
          required: true
          description: ID пользователя
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                address:
                  type: string
      responses:
        '200':
          description: Адрес пользователя успешно обновлен
        '400':
          description: Некорректные данные
        '404':
          description: Пользователь не найден
  /users/{id}/profile/external-links:
    get:
      tags:
        - Users
      summary: Получить список внешних ссылок
      operationId: getUserExternalLinks
      parameters:
        - name: id
          in: path
          required: true
          description: ID пользователя
          schema:
            type: integer
      responses:
        '200':
          description: Список внешних ссылок успешно получен
        '404':
          description: Пользователь не найден
    post:
      tags:
        - Users
      summary: Добавить внешнюю ссылку
      description: Добавляет информацию о стороннем ресурсе в профиль
      operationId: addExternalLink
      parameters:
        - name: id
          in: path
          required: true
          description: ID пользователя
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                platformName:
                  type: string
                  description: Название платформы (например, Telegram, GitHub, LinkedIn)
                  example: Telegram
                url:
                  type: string
                  description: Ссылка на профиль
                  example: https://t.me/username
      responses:
        '201':
          description: Ссылка успешно добавлена
        '400':
          description: Некорректные данные (например, невалидный URL)
        '404':
          description: Пользователь не найден
  /users/{id}/profile/external-links/{linkId}:
    delete:
      tags:
        - Users
      summary: Удалить внешнюю ссылку
      operationId: deleteExternalLink
      parameters:
        - name: id
          in: path
          required: true
          description: ID пользователя
          schema:
            type: integer
        - name: linkId
          in: path
          required: true
          description: ID ссылки, которую нужно удалить
          schema:
            type: integer
      responses:
        '200':
          description: Ссылка успешно удалена
        '404':
          description: Ссылка или пользователь не найдены
  /users/{userId}/profile/view:
    get:
      tags:
        - Users
      summary: Посмотреть профиль другого пользователя
      operationId: viewUserProfile
      parameters:
        - name: userId
          in: path
          required: true
          description: ID пользователя, чей профиль нужно просмотреть
          schema:
            type: integer
      responses:
        '200':
          description: Профиль пользователя успешно получен
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/User'
        '404':
          description: Пользователь не найден
        '403':
          description: У пользователя нет прав для просмотра этого профиля (например, если профиль скрыт)
    post:
      tags:
        - Users
      summary: Выполнить действия с профилем (заблокировать, добавить в друзья, общие комнаты)
      operationId: userProfileActions
      parameters:
        - name: userId
          in: path
          required: true
          description: ID текущего пользователя
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                action:
                  type: string
                  enum:
                    - block
                    - addFriend
                    - getCommonRooms
                  description: Действие, которое пользователь хочет выполнить
                targetUserId:
                  type: integer
                  description: ID другого пользователя (для блокировки или добавления в друзья)
      responses:
        '200':
          description: Действие успешно выполнено
        '400':
          description: Некорректные данные
        '404':
          description: Пользователь не найден
        '409':
          description: Ошибка , если пользователь уже заблокирован или добавлен в дрзья
  /users/{userId}/profile/view/friend-request:
    post:
      tags:
        - Users
      summary: Принять запрос в друзья
      operationId: acceptFriendRequest
      parameters:
        - name: userId
          in: path
          required: true
          description: ID текущего пользователя
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                targetUserId:
                  type: integer
                  description: ID пользователя, который отправил запрос в друзья
      responses:
        '200':
          description: Запрос в друзья успешно принят
        '400':
          description: Некорректные данные
        '404':
          description: Пользователь не найден
        '409':
          description: Запрос уже принят или отклонен
  /users/{userId}/profile/view/friend-request/decline:
    post:
      tags:
        - Users
      summary: Отклонить запрос в друзья
      operationId: declineFriendRequest
      parameters:
        - name: userId
          in: path
          required: true
          description: ID текущего пользователя
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                targetUserId:
                  type: integer
                  description: ID пользователя, который отправил запрос в друзья
      responses:
        '200':
          description: Запрос в друзья успешно отклонен
        '400':
          description: Некорректные данные
        '404':
          description: Пользователь не найден
        '409':
          description: Запрос уже принят или отклонен
  /users/{userId}/rooms:
    get:
      tags:
        - Users
      summary: Получить список своих комнат
      operationId: getUserRooms
      parameters:
        - name: userId
          in: path
          required: true
          description: ID пользователя, чьи комнаты нужно просмотреть
          schema:
            type: integer
      responses:
        '200':
          description: Список комнат пользователя
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Room'
        '404':
          description: Пользователь не найден или у пользователя нет комнат
  /users/{userId}/rooms/filter:
    get:
      tags:
        - Users
      summary: Фильтровать комнаты по времени или важности
      operationId: filterRooms
      parameters:
        - name: userId
          in: path
          required: true
          description: ID пользователя, чьи комнаты нужно фильтровать
          schema:
            type: integer
        - name: filterBy
          in: query
          required: true
          description: Критерий фильтрации (по времени или важности)
          schema:
            type: string
            enum:
              - time
              - importance
        - name: order
          in: query
          required: false
          description: Порядок сортировки (по возрастанию или убыванию)
          schema:
            type: string
            enum:
              - ascending
              - descending
      responses:
        '200':
          description: Список отфильтрованных комнат
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Room'
        '404':
          description: Комнаты не найдены
  /users/{userId}/rooms/folder:
    post:
      tags:
        - Users
      summary: Создать папку для комнат
      operationId: createRoomFolder
      parameters:
        - name: userId
          in: path
          required: true
          description: ID пользователя, который создаёт папку
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                folderName:
                  type: string
                  description: Название папки для комнат
      responses:
        '201':
          description: Папка успешно создана
        '400':
          description: Некорректные данные
        '404':
          description: Пользователь не найден
  /users/{userId}/rooms/folder/{folderId}/delete:
    delete:
      tags:
        - Users
      summary: Удалить папку для комнат
      operationId: deleteRoomFolder
      parameters:
        - name: userId
          in: path
          required: true
          description: ID пользователя, который удаляет папку
          schema:
            type: integer
        - name: folderId
          in: path
          required: true
          description: ID папки, которую нужно удалить
          schema:
            type: integer
      responses:
        '200':
          description: Папка успешно удалена
        '400':
          description: Некорректные данные (например, папка не пуста)
        '404':
          description: Папка или пользователь не найдены
  /users/{userId}/rooms/folder/{folderId}/update:
    put:
      tags:
        - Users
      summary: Изменить название папки для комнат
      operationId: updateRoomFolder
      parameters:
        - name: userId
          in: path
          required: true
          description: ID пользователя, который изменяет папку
          schema:
            type: integer
        - name: folderId
          in: path
          required: true
          description: ID папки, которую нужно изменить
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                folderName:
                  type: string
                  description: Новое название папки
      responses:
        '200':
          description: Папка успешно изменена
        '400':
          description: Некорректные данные
        '404':
          description: Папка или пользователь не найдены
  /users/{userId}/rooms/{roomId}:
    get:
      tags:
        - Users
      summary: Просмотреть информацию о конкретной комнате
      operationId: getRoomDetails
      parameters:
        - name: userId
          in: path
          required: true
          description: ID пользователя, чьи комнаты нужно просмотреть
          schema:
            type: integer
        - name: roomId
          in: path
          required: true
          description: ID комнаты, которую пользователь хочет просмотреть
          schema:
            type: integer
      responses:
        '200':
          description: Информация о комнате успешно получена
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Room'
        '404':
          description: Комната или пользователь не найдены
        '403':
          description: Пользователь не имеет доступа к этой комнате
  /users/{userId}/settings:
    get:
      tags:
        - Users
      summary: Получить текущие настройки пользователя
      operationId: getUserSettings
      parameters:
        - name: userId
          in: path
          required: true
          description: ID пользователя, чьи настройки нужно просмотреть
          schema:
            type: integer
      responses:
        '200':
          description: Текущие настройки пользователя
          content:
            application/json:
              schema:
                type: object
                properties:
                  notificationsEnabled:
                    type: boolean
                    description: Включены ли уведомления
                  privacy:
                    type: object
                    properties:
                      profileVisibility:
                        type: string
                        enum:
                          - public
                          - private
                          - friends
                        description: Кто может видеть профиль
                  language:
                    type: string
                    description: Язык интерфейса
        '404':
          description: Пользователь не найден
  /users/{userId}/settings/notifications:
    get:
      tags:
        - Users
      summary: Получить текущие настройки уведомлений
      operationId: getUserNotificationSettings
      parameters:
        - name: userId
          in: path
          required: true
          description: ID пользователя
          schema:
            type: integer
      responses:
        '200':
          description: Текущие настройки уведомлений
          content:
            application/json:
              schema:
                type: object
                properties:
                  roomInvitationNotifications:
                    type: boolean
                    description: Уведомления о приглашениях в комнаты
                  messageNotifications:
                    type: boolean
                    description: Уведомления о новых сообщениях
                  roomRemovalNotifications:
                    type: boolean
                    description: Уведомления об удалении из комнаты
                  # Оставляем старые поля, если они нужны для совместимости, или заменяем новыми правилами
                  emailNotifications:
                    type: boolean
                  pushNotifications:
                    type: boolean
        '404':
          description: Пользователь не найден
    put:
      tags:
        - Users
      summary: Изменить общие правила отправки уведомлений
      description:
        Настройка глобальных правил: приглашения, сообщения, удаление из комнат
      operationId: updateUserNotificationSettings
      parameters:
        - name: userId
          in: path
          required: true
          description: ID пользователя
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                roomInvitationNotifications:
                  type: boolean
                  description: Включить/отключить уведомления о приглашениях в комнаты
                messageNotifications:
                  type: boolean
                  description: Включить/отключить уведомления о сообщениях
                roomRemovalNotifications:
                  type: boolean
                  description: Включить/отключить уведомления об удалении из комнат
      responses:
        '200':
          description: Общие настройки уведомлений успешно обновлены
        '400':
          description: Некорректные данные
        '404':
          description: Пользователь не найден
  /users/{userId}/settings/general-notifications:
    get:
      tags:
        - Users
      summary: Просмотр общих настроек уведомлений пользователя
      operationId: getGeneralNotificationSettings
      parameters:
        - name: userId
          in: path
          required: true
          description: ID пользователя, чьи общие настройки уведомлений нужно просмотреть
          schema:
            type: integer
      responses:
        '200':
          description: Общие настройки уведомлений успешно получены
          content:
            application/json:
              schema:
                type: object
                properties:
                  emailNotifications:
                    type: boolean
                    description: Включены ли email уведомления
                  pushNotifications:
                    type: boolean
                    description: Включены ли push уведомления
        '404':
          description: Пользователь не найден
    put:
      tags:
        - Users
      summary: Обновить общие настройки уведомлений пользователя
      operationId: updateGeneralNotificationSettings
      parameters:
        - name: userId
          in: path
          required: true
          description: ID пользователя, чьи общие настройки уведомлений нужно обновить
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                emailNotifications:
                  type: boolean
                  description: Включить или отключить email уведомления
                pushNotifications:
                  type: boolean
                  description: Включить или отключить push уведомления
      responses:
        '200':
          description: Общие настройки уведомлений успешно обновлены
        '400':
          description: Некорректные данные
        '404':
          description: Пользователь не найден
  /users/{userId}/settings/privacy-and-recommendations:
    get:
      tags:
        - Users
      summary: Просмотр политики конфиденциальности и настроек рекомендаций
      operationId: getPrivacyAndRecommendations
      parameters:
        - name: userId
          in: path
          required: true
          description: ID пользователя, чьи настройки конфиденциальности и рекомендации нужно просмотреть
          schema:
            type: integer
      responses:
        '200':
          description: Политика конфиденциальности и настройки рекомендаций успешно получены
          content:
            application/json:
              schema:
                type: object
                properties:
                  privacyPolicy:
                    type: string
                    description: Политика конфиденциальности
                  recommendationSettings:
                    type: object
                    properties:
                      roomRecommendations:
                        type: boolean
                        description: Включены ли рекомендации комнат
        '404':
          description: Пользователь не найден
  /users/{userId}/settings/room-recommendations:
    put:
      tags:
        - Users
      summary: Изменить настройки рекомендаций комнат
      operationId: updateRoomRecommendations
      parameters:
        - name: userId
          in: path
          required: true
          description: ID пользователя, чьи настройки рекомендаций нужно обновить
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                roomRecommendations:
                  type: boolean
                  description: Включить или выключить рекомендации комнат
      responses:
        '200':
          description: Настройки рекомендаций комнат успешно обновлены
        '400':
          description: Некорректные данные
        '404':
          description: Пользователь не найден
...
components:
  schemas:
...
    User:
      type: object
      properties:
        name:
          type: string
        email:
          type: string
        id:
          type: integer
        passwordHash:
          type: string
        rooms:
          type: array
          items:
            type: integer
        interests:
          type: array
          items:
            type: integer
        feedbacks:
          type: array
          items:
            type: integer
        feedbacksAuthor:
          type: array
          items:
            type: integer
        createdAt:
          type: string
          format: date-time
    Room:
      type: object
      properties:
        adminId:
          type: array
          items:
            type: integer
        id:
          type: integer
        description:
          type: string
        category:
          type: string
        collaborators:
          type: array
          items:
            type: integer
        interestType:
          type: integer
        geolocation:
          type: string
        createdAt:
          type: string
          format: date-time
        createdBy:
          type: integer
        maxCollaborators:
          type: integer
...